Make valstat-ish parse APIs
