+++ include/simdjson/document.h
@@ -13,6 +13,7 @@
 namespace simdjson {
 
 template<size_t max_depth> class document_iterator;
+class document_parser;
 
 class document {
 public:
@@ -27,6 +28,9 @@ public:
   document &operator=(const document &o) = delete;
 
   using iterator = document_iterator<DEFAULT_MAX_DEPTH>;
+  class parser;
+  class doc_result;
+  class doc_ref_result;
 
   //
   // Tell whether this document has been parsed, or is just empty.
@@ -43,8 +47,6 @@ public:
   WARN_UNUSED
   bool dump_raw_tape(std::ostream &os) const;
 
-  class parser;
-
   //
   // Parse a JSON document.
   //
@@ -53,35 +55,10 @@ public:
   //
   // Throws invalid_json if the JSON is invalid.
   //
-  static document parse(const uint8_t *buf, size_t len, bool realloc_if_needed = true);
-  static document parse(const char *buf, size_t len, bool realloc_if_needed = true) {
-      return parse((const uint8_t *)buf, len, realloc_if_needed);
-  }
-  static document parse(const std::string &s, bool realloc_if_needed = true) {
-      return parse(s.data(), s.length(), realloc_if_needed);
-  }
-  static document parse(const padded_string &s) {
-      return parse(s.data(), s.length(), false);
-  }
-
-  //
-  // Parse a JSON document.
-  //
-  // If you will be parsing more than one JSON document, it's recommended to create a
-  // document::parser object instead, keeping internal buffers around for efficiency reasons.
-  //
-  // Returns != SUCCESS if the JSON is invalid.
-  //
-  static WARN_UNUSED error_code try_parse(const uint8_t *buf, size_t len, document &dst, bool realloc_if_needed = true) noexcept;
-  static WARN_UNUSED error_code try_parse(const char *buf, size_t len, document &dst, bool realloc_if_needed = true) {
-      return try_parse((const uint8_t *)buf, len, dst, realloc_if_needed);
-  }
-  static WARN_UNUSED error_code try_parse(const std::string &s, document &dst, bool realloc_if_needed = true) {
-      return try_parse(s.data(), s.length(), dst, realloc_if_needed);
-  }
-  static WARN_UNUSED error_code try_parse(const padded_string &s, document &dst) {
-      return try_parse(s.data(), s.length(), dst, false);
-  }
+  static doc_result parse(const uint8_t *buf, size_t len, bool realloc_if_needed = true);
+  static doc_result parse(const char *buf, size_t len, bool realloc_if_needed = true);
+  static doc_result parse(const std::string &s, bool realloc_if_needed = true);
+  static doc_result parse(const padded_string &s);
 
   std::unique_ptr<uint64_t[]> tape;
   std::unique_ptr<uint8_t[]> string_buf;// should be at least byte_capacity
@@ -90,30 +67,46 @@ private:
   bool set_capacity(size_t len);
 };
 
+class document::doc_result {
+private:
+  doc_result(document &&_doc, error_code _error) : doc(std::move(_doc)), error(_error) { }
+  doc_result(document &&_doc) : doc(std::move(_doc)), error(SUCCESS) { }
+  doc_result(error_code _error) : doc(), error(_error) { }
+  friend class document;
+public:
+  ~doc_result()=default;
+
+  operator bool() noexcept { return error == SUCCESS; }
+  operator document() {
+    if (!*this) {
+      throw invalid_json(error);
+    }
+    return std::move(doc);
+  }
+  document doc;
+  error_code error;
+};
+
+class document::doc_ref_result {
+public:
+  doc_ref_result(document &_doc, error_code _error) : doc(_doc), error(_error) { }
+  ~doc_ref_result()=default;
+
+  operator bool() noexcept { return error == SUCCESS; }
+  operator document&() {
+    if (!*this) {
+      throw invalid_json(error);
+    }
+    return doc;
+  }
+  document& doc;
+  error_code error;
+};
+
+
 } // namespace simdjson
 
-#include "simdjson/document/parser.h"
-#include "simdjson/document/iterator.h"
-
-// Implementations
-namespace simdjson {
-
-inline WARN_UNUSED document document::parse(const uint8_t *buf, size_t len, bool realloc_if_needed) {
-  document::parser parser;
-  if (!parser.allocate_capacity(len)) {
-    throw invalid_json(parser.error = MEMALLOC);
-  }
-  return parser.parse_new(buf, len, realloc_if_needed);
-}
-
-inline WARN_UNUSED error_code document::try_parse(const uint8_t *buf, size_t len, document &dst, bool realloc_if_needed) noexcept {
-  document::parser parser;
-  if (!parser.allocate_capacity(len)) {
-    return parser.error = MEMALLOC;
-  }
-  return parser.try_parse_into(buf, len, dst, realloc_if_needed);
-}
-
-} // namespace simdjson
+#include "simdjson/document_parser.h"
+#include "simdjson/document_iterator.h"
 
 #endif // SIMDJSON_DOCUMENT_H
