+++ include/simdjson/parsedjsoniterator.h
@@ -1,6 +1,6 @@
 #ifndef SIMDJSON_PARSEDJSONITERATOR_H
 #define SIMDJSON_PARSEDJSONITERATOR_H
 
-#include "document/iterator.h"
+#include "document_iterator.h"
 
 #endif
