+++ src/CMakeLists.txt
@@ -28,7 +28,7 @@ set(SIMDJSON_SRC
   stage1_find_marks.cpp
   stage2_build_tape.cpp
   document.cpp
-  document/parser.cpp
+  document_parser.cpp
   error.cpp
 )
